move duckdb and duckdb extension into pixels-duckdb
