+++ cpp/include/pixels_extension.hpp
