+++ cpp/pixels-duckdb/PixelsScanFunction.cpp
